+++ test/my_type.dart
@@ -1,0 +1,25 @@
+// @license
+// Copyright (c) 2025 Dr. Gabriel Gatzsche
+//
+// Use of this source code is governed by terms that can be
+// found in the LICENSE file in the root of this package.
+
+// .............................................................................
+class MyTypeWithoutJson {
+  const MyTypeWithoutJson(this.x);
+  final int x;
+}
+
+// .............................................................................
+class MyType {
+  const MyType(this.x);
+  final int x;
+
+  Map<String, dynamic> toJson() {
+    return {'x': x};
+  }
+
+  factory MyType.fromJson(Map<String, dynamic> json) {
+    return MyType(json['x'] as int);
+  }
+}